requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.20.0
urllib3<2.2.0
certifi>=2024.2.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# LLM Providers
openai==1.6.1
anthropic==0.8.1
cohere==4.40.0
google-generativeai==0.3.2
replicate==0.22.0
together==1.0.1

# HTTP Clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
numpy==1.24.4
pandas==2.1.4

# Utilities
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Redis (Optional)
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1